Refactor article page into multiple components
