+++ pages/article.js
@@ -2,64 +2,16 @@ import React from 'react';
 import { connect } from 'react-redux';
 import { compose } from 'redux';
 import moment from 'moment';
-import Link from 'next/link';
 import Head from 'next/head';
+import { nl2br } from '../util/text';
 
 import app from '../components/App';
 import ArticleInfo from '../components/ArticleInfo';
 import ArticleItem from '../components/ArticleItem';
+import CurrentReplies from '../components/CurrentReplies';
+import RelatedReplies from '../components/RelatedReplies';
+import ReplyForm from '../components/ReplyForm';
 import { load, submitReply, connectReply } from '../redux/articleDetail';
-
-const TYPE_NAME = {
-  NOT_ARTICLE: '⚠️️ 非完整文章或訊息',
-  NOT_RUMOR: '⭕ 含有正確訊息',
-  RUMOR: '❗ 含有不實訊息',
-};
-
-const TYPE_DESC = {
-  NOT_ARTICLE: '這篇不是轉傳訊息或完整網路文章，如「這個要怎麼用」、或訊息的節錄片段。',
-  NOT_RUMOR: '轉傳訊息或網路文章有一部分內容查證屬實。',
-  RUMOR: '轉傳訊息或網路文章部分含有有一資訊。',
-};
-
-const TYPE_INSTRUCTION = {
-  NOT_ARTICLE: '請簡單說明您為何認為這不是完整文章：',
-  NOT_RUMOR: '請簡單說明他哪個部分是正確的，作為「資料來源」的導讀：',
-  RUMOR: '請簡單說明不實之處，作為「資料來源」的導讀：',
-};
-
-function shortenUrl(s, maxLength) {
-  try {
-    s = decodeURIComponent(s);
-  } catch (e) {
-    // Probably malformed URI components.
-    // Do nothing, just use original s
-  }
-  return s.length <= maxLength
-    ? s
-    : `${s.slice(0, maxLength / 2)}⋯${s.slice(-maxLength / 2)}`;
-}
-
-const urlRegExp = /(https?:\/\/\S+)/;
-function linkify(str, maxLength = 80) {
-  if (!str) return '';
-  return str
-    .split(urlRegExp)
-    .map(
-      (s, i) =>
-        s.match(urlRegExp)
-          ? <a key={`link${i}`} href={s}>{shortenUrl(s, maxLength)}</a>
-          : s
-    );
-}
-
-function nl2br(text = '') {
-  const sentences = text.split('\n');
-  if (sentences.length <= 1) return sentences.map(s => linkify(s));
-  return sentences.map((sentence, i) =>
-    <div style={{ minHeight: '1.6em' }} key={i}>{linkify(sentence)}</div>
-  );
-}
 
 function getRatingString(replyConnections) {
   const resultStrings = [];
@@ -103,6 +55,42 @@ export default compose(
     handleSubmit = reply =>
       this.props.submitReply({ ...reply, articleId: this.props.query.id });
 
+    getStructuredData = () => {
+      const { data } = this.props;
+      const article = data.get('article');
+      const ratingString = getRatingString(data.get('replyConnections'));
+      if (!ratingString) return null;
+
+      // Ref: https://developers.google.com/search/docs/data-types/factcheck
+      //
+
+      return {
+        '@context': 'http://schema.org',
+        '@type': 'ClaimReview',
+        datePublished: moment(article.get('updatedAt')).format('YYYY-MM-DD'),
+        url: `https://cofacts.g0v.tw/article/${article.get('id')}`,
+        itemReviewed: {
+          '@type': 'CreativeWork',
+          author: {
+            '@type': 'Organization',
+            name: 'Internet',
+          },
+        },
+        claimReviewed: article.get('text'),
+        author: {
+          '@type': 'Organization',
+          name: 'Cofacts editors',
+        },
+        reviewRating: {
+          '@type': 'Rating',
+          ratingValue: '-1',
+          bestRating: '-1',
+          worstRating: '-1',
+          alternateName: ratingString,
+        },
+      };
+    };
+
     render() {
       const { data, isLoading, isReplyLoading } = this.props;
 
@@ -118,39 +106,8 @@ export default compose(
       if (article === null) {
         return <div>Article not found.</div>;
       }
-      const ratingString = getRatingString(replyConnections);
 
-      // Ref: https://developers.google.com/search/docs/data-types/factcheck
-      //
-      const structuredData = ratingString
-        ? {
-            '@context': 'http://schema.org',
-            '@type': 'ClaimReview',
-            datePublished: moment(article.get('updatedAt')).format(
-              'YYYY-MM-DD'
-            ),
-            url: `https://cofacts.g0v.tw/article/${article.get('id')}`,
-            itemReviewed: {
-              '@type': 'CreativeWork',
-              author: {
-                '@type': 'Organization',
-                name: 'Internet',
-              },
-            },
-            claimReviewed: article.get('text'),
-            author: {
-              '@type': 'Organization',
-              name: 'Cofacts editors',
-            },
-            reviewRating: {
-              '@type': 'Rating',
-              ratingValue: '-1',
-              bestRating: '-1',
-              worstRating: '-1',
-              alternateName: ratingString,
-            },
-          }
-        : null;
+      const structuredData = this.getStructuredData();
 
       return (
         <div className="root">
@@ -177,19 +134,7 @@ export default compose(
 
           <section className="section">
             <h2>現有回應</h2>
-            {replyConnections.size
-              ? <ul className="items">
-                  {replyConnections.map(conn =>
-                    <ReplyItem
-                      key={conn.get('id')}
-                      id={conn.get('id')}
-                      reply={conn.get('reply')}
-                      connectionAuthor={conn.get('user')}
-                      feedbackCount={conn.get('feedbackCount')}
-                    />
-                  )}
-                </ul>
-              : <p>目前尚無回應</p>}
+            <CurrentReplies replyConnections={replyConnections} />
           </section>
 
           <section className="section">
@@ -198,20 +143,11 @@ export default compose(
           </section>
 
           <section className="section">
-            <h2>相關文章的回應</h2>
-            {relatedReplies.size
-              ? <ul className="items">
-                  {relatedReplies.map(reply =>
-                    <RelatedReplyItem
-                      key={`${reply.get('id')}-${reply.get('articleId')}`}
-                      reply={reply}
-                      articleId={reply.get('articleId')}
-                      onConnect={this.handleConnect}
-                    />
-                  )}
-                </ul>
-              : <p>目前沒有相關的回應</p>}
-
+            <h2>相關文章回應</h2>
+            <RelatedReplies
+              onConnect={this.handleConnect}
+              relatedReplies={relatedReplies}
+            />
           </section>
 
           {relatedArticles.size
@@ -259,366 +195,10 @@ export default compose(
   }
 );
 
-function ReplyItem({ reply, connectionAuthor, feedbackCount }) {
-  const replyVersion = reply.getIn(['versions', 0]);
-  const createdAt = moment(replyVersion.get('createdAt'));
-  return (
-    <li className="root">
-      <header className="section">
-        {connectionAuthor ? connectionAuthor.get('name') : '有人'}
-        標記此篇為：<strong title={TYPE_DESC[replyVersion.get('type')]}>
-          {TYPE_NAME[replyVersion.get('type')]}
-        </strong>
-      </header>
-      <section className="section">
-        <h3>理由</h3>
-        <ExpandableText>{replyVersion.get('text')}</ExpandableText>
-      </section>
-      <section className="section">
-        <h3>出處</h3>
-        {replyVersion.get('reference')
-          ? nl2br(replyVersion.get('reference'))
-          : '⚠️️ 此回應沒有出處，請自行斟酌回應真實性。'}
-      </section>
-      <footer>
-        <span title={createdAt.format('lll')}>{createdAt.fromNow()}</span>
-        <span title="Coming soon!">・{feedbackCount} 人評價了這則回應</span>
-      </footer>
-
-      <style jsx>{`
-        .root {
-          padding: 24px;
-          border: 1px solid #ccc;
-          border-top: 0;
-          &:first-child {
-            border-top: 1px solid #ccc;
-          }
-          &:hover {
-            background: rgba(0, 0, 0, .05);
-          }
-        }
-        h3 {
-          margin: 0;
-        }
-        .section {
-          padding-bottom: 8px;
-          margin-bottom: 8px;
-          border-bottom: 1px dotted rgba(0, 0, 0, .2);
-        }
-      `}</style>
-    </li>
-  );
-}
-
-function RelatedReplyItem({ reply, articleId, onConnect }) {
-  const replyVersion = reply.getIn(['versions', 0]);
-  const createdAt = moment(replyVersion.get('createdAt'));
-  return (
-    <li className="root">
-      <header className="section">
-        <Link href={`/article/?id=${articleId}`} as={`/article/${articleId}`}>
-          <a>
-            其他文章
-          </a>
-        </Link>
-        被標記為
-        ：<strong title={TYPE_DESC[replyVersion.get('type')]}>
-          {TYPE_NAME[replyVersion.get('type')]}
-        </strong>
-      </header>
-      <section className="section">
-        <ExpandableText>{replyVersion.get('text')}</ExpandableText>
-      </section>
-      <footer>
-        <span title={createdAt.format('lll')}>{createdAt.fromNow()}</span>
-        ・<button type="button" value={reply.get('id')} onClick={onConnect}>
-          將這份回應加進此文章的回應
-        </button>
-      </footer>
-
-      <style jsx>{`
-        .root {
-          padding: 24px;
-          border: 1px solid #ccc;
-          border-top: 0;
-          &:first-child {
-            border-top: 1px solid #ccc;
-          }
-          &:hover {
-            background: rgba(0, 0, 0, .05);
-          }
-        }
-        h3 {
-          margin: 0;
-        }
-        .section {
-          padding-bottom: 8px;
-          margin-bottom: 8px;
-          border-bottom: 1px dotted rgba(0, 0, 0, .2);
-        }
-      `}</style>
-    </li>
-  );
-}
-
 function RelatedArticleItem({ article }) {
   return (
     <li>
       <ArticleItem article={article} />
     </li>
   );
-}
-
-const localStorage = typeof window === 'undefined' ? {} : window.localStorage;
-const formInitialState = {
-  replyType: 'NOT_ARTICLE',
-  reference: '',
-  text: '',
-};
-
-class ReplyForm extends React.PureComponent {
-  static defaultProps = {
-    onSubmit() {
-      return Promise.reject();
-    },
-    disabled: false,
-  };
-
-  constructor() {
-    super();
-    this.state = {
-      ...formInitialState,
-    };
-  }
-
-  componentDidMount() {
-    const { replyType, reference, text } = this.state;
-
-    // restore from localStorage if applicable.
-    // We don't do this in constructor to avoid server/client render mismatch.
-    //
-    this.setState({
-      replyType: localStorage.replyType || replyType,
-      reference: localStorage.reference || reference,
-      text: localStorage.text || text,
-    });
-  }
-
-  set(key, value) {
-    this.setState({ [key]: value });
-
-    // Backup to localStorage
-    if (typeof requestAnimationFrame !== 'undefined') {
-      requestAnimationFrame(() => (localStorage[key] = value));
-    }
-  }
-
-  handleTypeChange = ({ target: { value } }) => {
-    this.set('replyType', value);
-  };
-
-  handleTextChange = ({ target: { value } }) => {
-    this.set('text', value);
-  };
-
-  handleReferenceChange = ({ target: { value } }) => {
-    this.set('reference', value);
-  };
-
-  handleSubmit = e => {
-    e.preventDefault(); // prevent reload
-    if (this.props.disabled) return;
-    const { replyType, reference, text } = this.state;
-    this.props.onSubmit({ type: replyType, reference, text }).then(() => {
-      // Clean up localStorage on success
-      delete localStorage.replyType;
-      delete localStorage.reference;
-      delete localStorage.text;
-
-      this.setState(formInitialState);
-    });
-  };
-
-  handleSuggestionAdd = e => {
-    const suggestion = e.target.textContent;
-    this.set('text', `${this.state.text} ${suggestion}`);
-  };
-
-  render() {
-    const { replyType, text, reference } = this.state;
-    const { disabled } = this.props;
-
-    return (
-      <form onSubmit={this.handleSubmit}>
-        <p>
-          <select
-            name="type"
-            value={replyType}
-            onChange={this.handleTypeChange}
-          >
-            <option value="NOT_ARTICLE">{TYPE_NAME['NOT_ARTICLE']}</option>
-            <option value="NOT_RUMOR">{TYPE_NAME['NOT_RUMOR']}</option>
-            <option value="RUMOR">{TYPE_NAME['RUMOR']}</option>
-          </select>
-          <span>：{TYPE_DESC[replyType]}</span>
-        </p>
-
-        <div>
-          <label htmlFor="text">
-            {TYPE_INSTRUCTION[replyType]}
-          </label>
-          <br />
-          {replyType === 'NOT_ARTICLE'
-            ? <p>
-                常見原因 ——&nbsp;
-                <button
-                  className="suggestion"
-                  type="button"
-                  onClick={this.handleSuggestionAdd}
-                >
-                  長度太短，不像轉傳文章。
-                </button>
-                <button
-                  className="suggestion"
-                  type="button"
-                  onClick={this.handleSuggestionAdd}
-                >
-                  疑似為訊息之節錄片段，並非全文。
-                </button>
-                <button
-                  className="suggestion"
-                  type="button"
-                  onClick={this.handleSuggestionAdd}
-                >
-                  詢問句，非轉傳訊息。
-                </button>
-                <button
-                  className="suggestion"
-                  type="button"
-                  onClick={this.handleSuggestionAdd}
-                >
-                  測試用之無意義訊息。
-                </button>
-              </p>
-            : ''}
-          <textarea
-            required
-            id="text"
-            placeholder="140 字以內"
-            onChange={this.handleTextChange}
-            value={text}
-          />
-        </div>
-
-        {replyType === 'NOT_ARTICLE'
-          ? ''
-          : <p>
-              <label htmlFor="reference">
-                資料來源：
-              </label>
-              <br />
-              <textarea
-                required
-                id="reference"
-                placeholder="超連結與連結說明文字"
-                onChange={this.handleReferenceChange}
-                value={reference}
-              />
-            </p>}
-
-        <button className="submit" type="submit" disabled={disabled}>
-          送出回應
-        </button>
-        <span className="help">
-          不知道如何下手嗎？
-          <a
-            href="https://www.facebook.com/groups/cofacts/permalink/1959641497601003/"
-            target="_blank"
-            rel="noopener noreferrer"
-          >
-            Facebook 編輯求助區
-          </a>
-          歡迎您 :)
-        </span>
-
-        <style jsx>{`
-          textarea {
-            width: 100%;
-            height: 5em;
-          }
-          .suggestion {
-            background: transparent;
-            border: 1px solid #ccc;
-            border-radius: 4px;
-            padding: 8px;
-            margin: 0 8px 8px 0;
-            font-size: 12px;
-          }
-          .submit {
-            margin-right: 16px;
-          }
-          .help {
-            font-size: 12px;
-            font-style: italic;
-            display: inline-block;
-            color: #999;
-          }
-        `}</style>
-      </form>
-    );
-  }
-}
-
-class ExpandableText extends React.Component {
-  static defaultProps = {
-    children: '',
-    lines: 3,
-  };
-
-  constructor({ children }) {
-    super();
-
-    if (typeof children !== 'string') {
-      throw new Error('<ExpandableText> only accepts string children.');
-    }
-
-    this.state = {
-      isExpanded: false,
-    };
-  }
-
-  toggleExapnd = () => {
-    this.setState({ isExpanded: !this.state.isExpanded });
-  };
-
-  render() {
-    const { children, lines } = this.props;
-    const { isExpanded } = this.state;
-    const sentences = nl2br(children);
-
-    if (sentences.length <= lines) {
-      return (
-        <div>
-          {sentences}
-        </div>
-      );
-    }
-
-    return (
-      <div>
-        {isExpanded ? sentences : sentences.slice(0, lines)}
-
-        <button className="more" onClick={this.toggleExapnd}>
-          {isExpanded ? '隱藏全文' : '閱讀更多'}
-        </button>
-        <style jsx>{`
-          .more {
-            border: 0;
-            background: transparent;
-            text-decoration: underline;
-          }
-        `}</style>
-      </div>
-    );
-  }
 }
